Third failing test
Chose 5 = V as the next failing test case. This is actually an
interesting decision, in some ways 2 = II is a more obvious choice.
However, 5 = V is really a simpler case as II would involve more
decisions about the implementation such as whether to hardcode the "II"
string or somehow create it from two "I" strings