Refactoring towards sensible approach
This was quite a big refactoring step but it has moved me away from the
clearly silly solution of having every value map to its corresponding
roman numeral! It was clear that I would have to treat "II" as a case of
two "I"s and at that point a recursive solution seemed to suggest
itself.