Additional refactoring to use domain specific type
I didn't like the mixing of abstraction levels in the generate function.
It seemed confusing to be refering to keys and values where the key was
the value and the value was the roman numeral!
So I have introduced a new type (RomanNumeralMapItem) with properties of
Value and RomanNumeral to hopefully keep the generate function more
readable.