Complete
I decided it was not worthwhile to try and calculate values such as 4 =
IV and instead to keep those as seperate items in the map. The
additional complexity to the algorithm for those cases did not seem
worth it. Also, once I was confident that my approach was suitable for a
large number of items I decided to move to a single test asseting
against a collection of test data to avoid lots of duplication in the
test code.